Flip the verifiers and airdrop so claim doesnt need to construct a transaction
