+++ Anchor.toml
@@ -9,7 +9,7 @@ merkle_verifier = "4ibGmfZ6WU9qDc231sTRsTTHoDjQ1L6wxkrEAiEvKfLm"
 password_verifier = "EmsREpwoUtHnmg8aSCqmTFyfp71vnnFCdZozohcrZPeL"
 
 [provider]
-cluster = "localnet"
+cluster = "devnet"
 wallet = "~/.config/solana/id.json"
 
 [scripts]
