+++ programs/airdrop/src/errors.rs
@@ -2,8 +2,6 @@ use anchor_lang::prelude::*;
 
 #[error_code]
 pub enum ErrorCode {
-    #[msg("Verifier program in claim does not match configure")]
-    IncorrectVerifierProgram,
-    #[msg("Verifier state in claim does not match configure")]
-    IncorrectVerifierState,
+    #[msg("Verifier signer did not match expected")]
+    IncorrectVerifierSignature,
 }
