+++ programs/airdrop/src/instructions/configure.rs
@@ -5,7 +5,7 @@ use crate::State;
 use crate::VAULT_SEED;
 
 #[derive(Accounts)]
-#[instruction(state_seed: [u8; 32], verifier_instruction_prefix: [u8; 8])]
+#[instruction(state_seed: [u8; 32])]
 pub struct Configure<'info> {
     #[account(mut)]
     pub payer: Signer<'info>,
@@ -31,10 +31,8 @@ pub struct Configure<'info> {
     pub vault: Account<'info, TokenAccount>,
     pub mint: Box<Account<'info, Mint>>,
 
-    /// CHECK: Will be verified when used in cpi.
-    pub verifier_program: UncheckedAccount<'info>,
     /// CHECK: Will be verified in the verifier_program.
-    pub verifier_state: UncheckedAccount<'info>,
+    pub verifier_signature: AccountInfo<'info>,
 
     pub token_program: Program<'info, Token>,
     pub system_program: Program<'info, System>,
@@ -44,11 +42,9 @@ pub struct Configure<'info> {
 pub fn handle_configure(
     ctx: Context<Configure>,
     state_seed: [u8; 32],
-    verifier_instruction_prefix: [u8; 8],
 ) -> Result<()> {
-    ctx.accounts.state.verifier_program = ctx.accounts.verifier_program.key();
-    ctx.accounts.state.verifier_state = ctx.accounts.verifier_state.key();
-    ctx.accounts.state.verifier_instruction_prefix = verifier_instruction_prefix;
+    ctx.accounts.state.verifier_signature = ctx.accounts.verifier_signature.key();
+
     ctx.accounts.state.vault = ctx.accounts.vault.key();
     ctx.accounts.state.vault_bump = *ctx.bumps.get("vault").unwrap();
     ctx.accounts.state.close_authority = ctx.accounts.payer.key();
