+++ tests/utils/utils.ts
@@ -139,3 +139,12 @@ export const toBytes32Array = (b: Buffer): number[] => {
 
   return Array.from(buf);
 };
+
+export function toBeBytes(x: number) {
+  const y = Math.floor(x / 2 ** 32);
+  return Uint8Array.from(
+    [y, y << 8, y << 16, y << 24, x, x << 8, x << 16, x << 24].map(
+      (z) => z >>> 24,
+    ),
+  );
+}
